requirements.txt
keyboard
opencv-python
paho-mqtt
pyyaml
